update and unify test
